Move volatile database to application source tree. #3
