--- esp32/components/database/CMakeLists.txt
@@ -1,12 +1,0 @@
-set(SOURCES
-)
-
-if(ESP_PLATFORM)
-
-idf_component_register(SRCS ${SOURCES}
-                       INCLUDE_DIRS .
-                       REQUIRES main)
-
-else()
-
-endif()
